+++ bit/cli/calc_gc_per_seq.py
@@ -13,13 +13,13 @@ def main():
 
     parser = argparse.ArgumentParser(
         description=desc,
-        epilog="Ex. usage: bit-calc-gc-per-seq -i input.fasta -o output.tsv",
+        epilog="Ex. usage: `bit-calc-gc-per-seq -i input.fasta -o output.tsv`",
         formatter_class=CustomRichHelpFormatter,
         add_help=False
     )
 
-    required = parser.add_argument_group("REQUIRED PARAMETERS")
-    optional = parser.add_argument_group("OPTIONAL PARAMETERS")
+    required = parser.add_argument_group("Required Parameters")
+    optional = parser.add_argument_group("Optional Parameters")
     required.add_argument(
         "-i",
         "--input-fasta",