+++ bit/cli/kraken2_to_taxon_summaries.py
@@ -20,8 +20,8 @@ def main():
         add_help=False
     )
 
-    required = parser.add_argument_group("REQUIRED PARAMETERS")
-    optional = parser.add_argument_group("OPTIONAL PARAMETERS")
+    required = parser.add_argument_group("Required Parameters")
+    optional = parser.add_argument_group("Optional Parameters")
 
     required.add_argument(
         "-i",