+++ bit/cli/summarize_column.py
@@ -23,8 +23,8 @@ def build_parser():
         formatter_class=CustomRichHelpFormatter,
         add_help=False
     )
-    required = parser.add_argument_group("REQUIRED PARAMETERS")
-    optional = parser.add_argument_group("OPTIONAL PARAMETERS")
+    required = parser.add_argument_group("Required Parameters")
+    optional = parser.add_argument_group("Optional Parameters")
 
     required.add_argument(
         "-i",